検索機能完成

+++ app/Http/Controllers/PostController.php
@@ -12,8 +12,9 @@ class PostController extends Controller
         $posts = Post::select('posts.id as post_id', 'posts.user_id', 'users.name', 'posts.content', 'posts.game_id', 'posts.start_at')
         ->join('users','posts.user_id','=','users.id')
         ->get();
-        
-        return view("post/index",compact('posts'));
+        $user_id = Auth::id();
+
+        return view("post/index",compact('posts','user_id'));
     }
     
     public function store(Request $request){
@@ -34,20 +35,33 @@ class PostController extends Controller
     
     public function edit($post_id){
         $post = Post::where('id','=',$post_id)->first();
+        $user_id = Auth::id();
+        if($post->user_id != $user_id){
+            return redirect('/post');
+        }
         return view("post/edit",compact('post'));
     }
     
     public function update(Request $request,$post_id){
         $post = Post::where('id','=',$post_id)->first();
+        $user_id = Auth::id();
+        if($post->user_id != $user_id){
+            return redirect('/post');
+        }
         $post->content = $request ->content;
         $post->game_id = $request ->game_id;
         $post->start_at = $request ->start_at;
         $post->save();
+        
         return redirect('/post');
     }
 
     public function delete($post_id){
         $post = Post::where('id','=',$post_id)->first();
+        $user_id = Auth::id();
+        if($post->user_id != $user_id){
+            return redirect('/post');
+        }
         $post->delete();
 
         return redirect('/post');    
@@ -57,8 +71,11 @@ class PostController extends Controller
         // ユーザーが選択した、game_idを取得
         $game_id = $request->game_id;
         // postsテーブルのcontentカラムの中から、game_idが部分一致するものを取得
-        $posts = Post::where('content', 'like', "%$game_id%")->get();
-        // view側にpostsの配列を渡す
-        return view('post',compact('posts'));
+        $posts = Post::select('posts.id as post_id', 'posts.user_id', 'users.name', 'posts.content', 'posts.game_id', 'posts.start_at')
+        ->join('users','posts.user_id','=','users.id')
+        ->where('game_id', '=', $game_id)
+        ->get();        
+        $user_id = Auth::id();
+        return view('post/index',compact('posts','user_id'));
     }
 }
